11-06 update

+++ ISA-2Dproject/Assets/Scripts/WeaponScript.cs
@@ -2,7 +2,7 @@
 
 public class WeaponScript : MonoBehaviour
 {
-    public float weaponStateNumber;
+    public int weaponStateNumber;
     public float localStartTimeBtwAttacks;
     public float localAttackRange;
     public GameObject localWeaponText;
